fix(compilation): Resolve extensive compilation errors across all packages
- Fix LLM package: BatchProcessor interface, dynamic context manager, performance client
- Fix Nephio Porch: lifecycle manager, workflow engine, package revision
- Fix Security CA: x509 key usage, crypto interfaces, validation framework
- Fix Performance: stat functions, memory pools, regression detection
- Fix O-RAN O2: OpenStack provider types, VMware compatibility
- Fix Testing: framework compatibility, validation suite conflicts
- Fix Auth & Monitoring: interface implementations, type conversions

+++ pkg/performance/goroutine_pools.go
@@ -140,7 +140,7 @@ type PoolMetrics struct {
 	Throughput         float64 // tasks per second
 	CPUUtilization     float64
 	MemoryUsage        int64
-	GoroutineCount     int
+	GoroutineCount     int64
 	ScalingEvents      int64
 	PreemptionEvents   int64
 	AffinityViolations int64
